Moved the predicate generation code to ColumnTableScan Added stats to keep track of scanned cached batches
